Add httpdate OpenRC init script
The usual mechanism for synchronizing the local system's time with the
external world is by means of NTP. However, there might be cases where
NTP access is blocked by a firewall or such a service is not available
for other reasons.
In such cases it seems reasonable to fall back to using the time stamp
of an HTTP request of a well known server to set the (rough) time. This
change introduces the httpdate init script which attempts to synchronize
the local time using the date from an HTTP request.