Force using of /bin/date
On certain environment we have seen the update of the system date fail
with an error that the given time format is invalid. The exact cause of
this problem is still unclear but it seems that 'date' refers to
something else than /bin/date in those cases and that this entity is
less liberal with respect to date formats.
This change forces the usage of /bin/date which makes those errors
disappear.