Retry time synchronization upto five times
Unfortunately the init functionality on Gentoo leaves some problems
unresolved. In particular, although we rely on net being started there
is no guarantee that we have received an IP address and DNS information
by the time this init script is run. We have seen this problem
repeatedly on the raspberry pi.
With this change we retry five times, with a break of three seconds
between retries. This process seems to allow enough time for the
network to become fully available, although no guarantee is possible, of
course.