Special case time setting for busybox' date
Busybox' version of date(1) does not recognize the standard HTTP
reported date. So in order to convince it to still set the time, we have
to tell it the format in which the time is in. What's worse, it uses the
special switch -D for setting the format, which no other implementation
of date seemingly supports. To that end, we have to fully special case
the invocation for busybox.