Replace usage of wget with curl
As it turns out wget, which we ultimately use to retrieve the date of
google.com, has a dependency on Perl. That is utterly unacceptable for
at least one of the use cases for this script.
To fix this problem, this change replaces the usage of curl with wget,
which is slimmer in many respects.